Move codecs helpers to utils/codecs
